Local Multiplayer Re-write
Broke the VR player mode 'rep-notify' into three different function instead to allow local multiplayer to work.  System now allows for multiple player pawns to be placed in a scene.  Changed the enter actor logic to warp the avatar to the actor if there is a scene locator and multiple player pawns are present (this allows the user to preserve real world spatial relations).