VRMocap_MovementActor
This update brings a mostly functional character controller that allows both physical locomotion and character movement locomotion using a capsule.  Under components, there is a blueprint for the VRMocap_MovementActor which can be copied and modified to have whatever skeletal mesh you want (even the Metahumans).  The mocap actor component and the meshes need to have the proper animation blueprints on them still however!