Update to 4.27  - Unknown Actor Type
Updated the system to UE4 4.27

Cleaned up Tracker Recording Rig's run animations, and parented this rig under the mocap actor forward, so the run cycles should line up better.

Fixed the Unknown Actor Type:
Added a new actor profile that will automatically find the closest bone to your steamvr tracked points.  This set-up is fairly stable, but I have noticed crashes.  To get the best performance. do the following:

Right click on the VRMocap_Unknown_AnimBP and retarget to the skeleton of the character you want to use.  Make sure your mesh is using this animation bp.

Make sure that you enter in the pelvis name to the Mocap Actor Component.  If you want finger motion, put a 1 or -1 onto the axis that the thumb and fingers should bend.

In the VRMocap_Unknown_ControlRig, delete whatever bone hierarchy is in there currently, and add the skeleton of whatever character you want to motion capture.

Currently the movement actor does not work with the unknown actor profile.