4.2.7 iPhone Only Mode
Added a bool on the VR Mocap Player Pawn so that a character can be controlled just with an iPhone.  After 'calibrating' the torso, neck and body will be animated by the head rotation data from live link face.

Added a ue4 mannequin to the project to prevent some of the errors on load time.