4.27 - Bug Fixes from physical update
I noticed that the Example Level in the VRMocap folder wasn't set up properly.  The actor was set to target a different player pawn index than was what was in the scene (something I probably did while testing local multiplayer) so to avoid confusion I made it match the player pawn that you autospawn into.

I noticed some weird glitches with entering and exiting ragdolling, so I changed that logic.  Things should work better now.  The biggest issue I found while testing is turning on physical animations WHILE in a ragdoll state often makes it so the character will be rotated to the floor.