-4.27  Player Pawn Interfaces Refactor, No Chest Tracker, Control Rig Updates
The movement actor has a new mode that locomotes in relation to the VRMocap_Spectator Camera.

Revised how the 'no chest tracker' logic works: should give much better results (although the Metahuman seems to be impossible to get right as the model scales from the set-up phase to the forward processing phase)

Added a lot of new logic to the metahuman control rig, included a new kind of clav estimation.  In most of my tests it looks better, and I might port the logic to the CC3 character and mannquin characters.  Essentially, a clav controller points at the hand, and when the arm is above a certain heigth  we rotate the clav bone slightly to match.

Refactored the player pawn to not use casts anymore.  This should prevent errors if you aren't using one of the supported skeletons in your project, and should make extending the system easier for custom models.  Right now, setting which trackers are active, the mocap control floats, entering/exiting actor events and the sending of tracker transform data are handled by sending the data to the actor through the VRMocap_Interface.

Worked on getting the unknown actor type to use the movement actor.  Still doesn't work, but it's closer?