4.27- Movement Actor No Feet Mode
Lots of work on the movement actor.  If you have no feet trackers listed, the system will attempt to combine your IRL and in game movement speed and play a walking animation based on the 8 way movement blendtree.  It's not perfect, but this allows you to remove 2 trackers from an actor, and can look decent in cowboy shots (MS framing).  Now if I could only figure out a good way to estimate the player's hip location, it would be possible to use this system with 3 VR devices.

I refined the operation of camera based movement.  I created a system to rotate the character towards their destination without using the movement actor system (this prevented popping when the thumbstick was released, although this still occasionally happens with the UE4 mannequin, and will always happen if you only do a short thumbstick move).  I have an idea to create a 'camera cut' event that would allow players to use the same input direction unless the stick is released or drastically changed in direction, although this hasn't been implemented.

I couldn't figure out why calibrating the movement actor with any rotation caused problems, so I forced its rotation to 0 using a construction script.

 Adjusted Metahuman head IK.

Wrote some additional code in the mocap actor component that should exclude skeletal meshes that don't have certain bone names.  I noticed sometimes if a character had a prop, the playerpawn would add the item as the mocap actor if it was a skeletal mesh.

Made the UE4 mannequin control rig more like the Metahuman and CC3 control rig.