4.27 - No Feet IK, External Actors
No feet mode mode now has IK.  Sometimes it gets stuck, and is pretty garbage, but at least there is less foot sliding now.

The movement actor can now cast its values to an external actor.  I noticed that when recording this actor, it gives you a lot of unwanted animation tracks and bad placement of the model.  This should allow you to put in a the same actor as a traditional mocap actor set up (facing x, with the animation blueprint and mocap actor component) and you can record the data on that character.

Added a new metahuman face with UE4 mannequin body type.  You attach the head before runtime the same as with the CC3 actor.

Revised the CC3 control rig to use the Metahuman clav controller.  Before, I noticed in extremely long scenes, the old control rig would cause the joint to gradually move downwards.  Made the UE4 Mannequin use the same clav system as the metahuman and CC3 character (basically there is a controller aiming at the hand that helps move the clav to align with where the hand goes).

Fixed an issue with VTuber mode where it just scaled down to 0,0,0.

Fixed a bug with the Editor Utility Widget where it wasn't loading calibration data.

I did some other stuff too, but I forgot.