UE5 P1 - Control Rig Fix and general maintenance
Changed some of the bone stretching behavior, and reverted from the new full body IK system for the spine to the old deprecated one (this one just seems to work better for me, and I hope it never leaves, haha).

Changed the amount of stretching the arms and legs can do on the basic IK node
Removed the public variable for 'tracker controlled' on the spectator camera: now if the live link tracker field has a live link subject, it will automatically track.  To make it so the camera always tracks (even during calibration) then check on the 'always track camera).

Changed the default smoothing amount on the tracker controlled camera from 20 to 12.  This should remove some jittery movements (lower numbers are smoother, but lag behind your actual move more).

Fixed a bug where iphone only mode wasn't working on CC3/Mannequin Metahumans.

Noticed the anim bp live link face on mannequins was using the wrong remap asset.

Made potato mode also switch off lighting, so that should allow for better framerates when capturing metahumans.

The start of a new actor type is in the project now, but not ready yet.  I deleted the old debugging humanoid model, and make a new one for the 'humanoid' actor type.  I'm working on a control rig that automatically finds the closest bone to a rig you fit to your actor.  I think it still needs axis information for where limbs are pointing, but it's nearly ready to test out.