UE5.0.1 Sequencer Slots
OK, sorry for the lack of updates.  UE5 completely wrecked my live link XR on my main PC, and after several reinstalls I can't fix the issue.  Make sure you do a Windows system backup before moving any important VR Mocap projects to 5.0.

I've been working on little things here and there for the last month, and have found it hard to keep track of what has changed... so let me know if anything isn't working.

The big update is the animation blueprints now support slots for 'face' 'hand_l' and 'hand_r' meaning you can either use a pose or a previously captured mocap file and override the character's face or hands.  To do this, it is REQUIRED that you add animations to an actor that IS ALREADY IN THE SCENE (spawning them in with the lightning bolt icon does not seem to activate their animation blueprint, and even choosing to use the class's animation blueprint specifically in sequencer...)

I replaced the UE4 mannequin control rig spine with the new FBIK.  It looked better there than it did for the other skeletons, so all the other skeletons are the same.

In other news, I am struggling to make the 'HUMANOID' actor profile type, which intends to dynamically build a rig with bone names you give it or that it gets during it's construction script.  For some reason, the control rig doesn't seem to animate the character, which is a bummer.  I think I could use some help with that.