UE5.0.1 - Audio Based Mouth Movements
Made a new bool on the player pawn that creates a new 'AudioCapture_BP' that will get the amplitude of your current sound input and use that to drive the AR Kit 'jawOpen' morph target.  This currently doesn't work with metahumans, and I can't find the magic parameter for this...

Added a DB meter in the editor widget so you can see the amplitude of your current input.  Hopefully this can help serve as a warning when you are recording to be a remind if you are clipping or potentially using the wrong audio input (I often record audio with the body motion, so it's a nice visual look at what is going on).

Cleaned up some directories by moving items around.