UE5.0.2 - UE5 Mannequin and IK Retargeting update
Added a new actor profile type: UE5 Mannequin!  This uses the Metahuman control rig, and I think it could be tweaked somewhat, but seeing as this is the new standard I thought it would be good to put in the system!

Added the set up for a new system that uses the IK retargeting set up to transfer data from a UE5 mannequin to any other mesh with an IK retargeting asset made for it.  To use it, you will have to retarget this animbp to your mesh, and set up a new IK retargeting node that goes from UE5 mannequin to your character, and update the IK retageting node in the blueprint with this new remapping asset.  Anyways, this really opens up the amount of skeletons you can work with.  It works, but at what cost?

Up on the d-pad will trigger the frame actor command on the camera.

Added a new bool that defaults to on when using a scene locator: this will force the placespace to that location.  Also, the player pawn will be made a child of the locator object, so this object can be moved around at runtime to place your real world space exactly where you want it in the game world.

Added a bone compression preset for take recorder that is less destructive than the default.  You have to replace the default one located in Engine/Content/Animation/DefaultRecorderBoneCompression

Made the control rigs for UE4 Mannequin and CC3 update the animation on the characters IK bones based on the tracker position.

Changed the metahuman control rig (which the UE5 mannequin uses as well) to no longer use the clav calculation I set up.  One of the IK nodes must have updated, and it knocked some prior settings.  In any case, these characters should look better now.

Modified the curve of CC3 character's smiles to be 80% strength.

Deprecated the slot functionality on the cc3 and mannequin.  This was intended to allow people to drive different animations on different bones/blendshapes, but ultimately it doesn't work with the renderer so I disconnected this feature.  Results do work in editor/at runtime however.

Changed the logic flow of the autofocus system, so it shouldn't go out of focus when entering POV mode.