UE5.0.3 - UE5 Manny Full Body IK Solver
Put a lot of work into the UE5 Manny full body IK solver.  The chest/neck rotations should be better, and there is additional hip movement from the legs.  Also changed the hand controls to use the VRMocap Hand Control Rig, which will give smoother values and a nice breathing motion to the hands.

Changed the directory of the UE5_Manny BP so it is easier to find incase someone wants to manually add it to a scene.

Revised the metahuman animbp to use the new hand set up as well, although it is still using the old solver.  You should be able to swap out the metahuman body control rig for the UE5 manny control rig that has been added to this submit to see if that gives you better performance.

Changed the spawning behavior of the IK retargeting actor to appear right where the old character mesh was located.

Changed the way that the system finds the UE5 Manny mesh in the MocapActor_Component.  Hopefully this will avoid conflicts when the system attempts to find the mesh in the movement actor.

Revised the movement actor to come with two meshes and a bool for IK retargeting.  Turning this bool on will hide the UE5 manny mesh after calibration and reveal the mesh parented underneath.  If you set up this character with an IK retargeting animation blueprint and change the IK retargeting asset in the animation graph, then you can use whatever mesh you want with the movement actor.

Added a fake pelvis tracker that will kick in if you delete out the pelvis tracker live link device name on the player pawn.  Don't expect any miracles though!