UE5.0.3 - Backwards Solvers and Control Rig Refactor
I redid the Metahuman and UE5 Manny control rigs to not use a fake head controller, which seemed to give better head motion when low to the ground.  I might need to bring some of this fake head behavior back depending on popular opinion.

Added backwards solvers for all the main control rigs.  This should add a ton of power when cleaning up your animations.

Removed a lot of older controls from the control rigs so that using them with backwards solves would be easier.

Added a way to shrug when using metahumans or UE5 mannequins: the system will use the elbow position to estimate where the shoulder is, so make sure those trackers go on your upper arm.  Thanks to Fofamit for this suggestion.

There was a bug where ue4 animation blueprints got steam finger curls only from the left hand.

Deleted the VRMocap folder content which seemed to persist in the github repo after I refactored.

Added a janky VR mirror.  Don't look at it too closely or you might enter a dimensional vortex of pain and suffering... or just throw up because I couldn't figure it out.