UE5.0.3 - Spawning System
Spent a lot of work getting a spawning system in: there is now an event that takes an blueprint class (your actor that has been set up with a mocap actor component and relevant anim bp) and calling it will delete the current actor and spawn you into the new actor using the latest saved calibration data.  If things look weird, just calibrate that character again on the spot.

The biggest reason this update took so long is I was fighting a bug with the IK retargeting in 5.0.  Hopefully this will be resolved in 5.1, but I had to do some things I'm not proud of, and that no person should have to see.

I added an 'auto arm calibrator' that moves the arms to match your controllers, and while it's not perfect I have found it so useful that I have enabled it by default.  You can turn this feature off in the mocap actor component.

Added the MMVR Avatar (the mocap skeleton character) to the projects so you can make little animations with them.  in the editor utility widget, the 'load actors' buttons will actually spawn in new characters.

Lots of little fixes and clean up: it's been months, so I assume most systems have been touched in some way.