UE5.1 - Upgrade To 5.1
NOTE:  SteamVR is deprecated in 5.1, however at the current moment it is the only reliable way that Live Link XR tracked devices show up in Live Link (and also I really like the SteamVR finger curl system), so for now continue using the SteamVR plugin and NOT OpenXR.  There may come a point in which the SteamVR plugin will need to be downloaded from Valve's github and manually installed to your Unreal Engine.

In the player pawn, I added the enhanced input events, however these seem to only fire when the game is run in VR mode.  They are in the graph in case the old input system is deprecated fully (which unfortunately sounds likely).

Added Oculus controllers to the input.ini file, as well as grip events.

Added enhanced input assets, as well as some haptic stuff I don't understand/doesn't work (I'm going to blame OpenXR and say it is their fault because they have ruined all that is good and holy in how I use Unreal Engine).

Updated the camera system to look at if you are using the HMD to capture your scene and then update which components will be deleted.  This should fix any black screens on the spectator camera when running VR Mode.  Also, updated that render texture to be 1920x1080.

Updated the Unknown Body Control Rig to use the new Full Body IK set up.  As of this release, that mode still does not work however.