UE5.1 - POV Camera and Springy Goals
Added a POV Camera actor with accompanying Editor Utility Widget.  On play, the camera will get attached to your skeletal mesh, giving you a second render view to show you that character's perspective.  This includes a Teleprompter widget, that will take in a line and actor name and briefly display it on screen when 'Event_ChangeText' is called.

Want some more spring in your step!  Use the Springy Goals float 0-1 in the MMVR_ActorComponent or the Editor Utility Widget to add some spring effectors to your ik goals.

Changed Chest Control to Spine 04 on UE5 Manny.  I think the solve looks better now, but tell me what you think!

Reskined the UE5 manny model, so deformations are easier to see on some of the additional torso bones.  Some of the weight painting is a little sharper than I wanted, but thems the breaks.

Set limits on hand control rig so that the control points can't be lifted out of the scene, and set up reverse solving.

General Cleanup and renaming.  Changed some material properties so they should look better in all scenes.