UE5.1.1 - Take Recorder, Pt. 1
Cleaned up the MMVR Actor Component to have less casts, and have more reliable flow.

Added take recorder functionality so that the user can set up a take and record a take from within blueprints!  This could be extremely useful for people creating cutscenes, and I already see a lot of potential here.  Check it out on the Editor Utility Widget.

Added a way to send teleprompter messages from the Editor Utility Widget.

Changed the inputs so that holding both B buttons cause a ragdoll, and holding B on the controllers creates a detach event.  Right now, holding the B button on the right controller will detach the spectator camera from whatever actor is holding it, which can allow the user to make some fun angles.

If the player is not using a movement actor, they can use the joysticks to move themselves around the X and Y plane.  This should help for adjusting start positions if you aren't where you need to be.

Removed some old files that were no longer in use.  I will have to remove a lot more to keep things tidy: let me know if you need any of the debugging skeletons!