UE5.1.1 - IK Retargeting Refactor
Did a refactor of the IK retargeting path.  Before the system would spawn in an actor as an IK retargeting source, but I've included that mesh inside the player pawn  This means that the mocap actor property in the player pawn will always remain your mocap target.  This should simplify the system, and make it more robust, although currently this path isn't working with the movement actor.  For the movement actor, replace the skeletal mesh that is a child of the main character mesh and make the public bool for IK retargeting on the movement actor true.

Removed some elements I was experimenting with to get a drag and drop interface in the editor utility widget.  I could use help with this still if anyone has expertise there.

Added an MMVR_DataTableObject for holding script data.  This script can be put into the editor utility widget, and used in the Teleprompter field.

Removed the clav movement component responsible for adding shoulder shrugs in the UE5 manny face control rig because it was causing weird arm flipping.  I would love to figure out how to get the elbow trackers to also offset the clav rotation if anyone has any ideas.

Added the ability to offset the location and rotation of a mocap actor while performing, as well as to change the field of view on cameras.