UE5.1.1 - CC4/CC3 Plus and Lighting Rig
Added a new anim bp for CC4/CC3 Plus characters that use the extended face shapes profile.  This should work with wrinkles, although occasionally the shader breaks in my testing.  Reguardless, this is a great new step for people using the new blendshapes.  If you want to use the old CC3 behavior, I still have that anim bp, although there is a 'DEP_' prefix added to it to avoid confusion.

Added a new lighting rig that SHOULD detect cuts in sequencer and fast camera motion.  The idea is that it will always point at the active camera rendering the scene, so your lighting set up should remain relative to that and the actors position on the cut.  If you add in a preset name, you can save and load lighting set ups, so you can make a library of moods and looks for your cinematics and vtuber set ups.  I could use some help figuring out how to load all your presets in a drop down list, as well as flushing out a control scheme I made using utility widgets if anyone is out there!

Basic reorganization and clean up.