UE5.1.1 - Light Rig Hot Fix
I found that some buttons weren't updating values in the light parameters ui, so I fixed that.  Unfortunately the 'is light below' is a little sticky and doesn't work 100% as expected, but I'm considering this an acceptable loss all things considered.  Also, there was a bug where lighting set ups kept getting added to the master list, but that is fixed now.