UE5.1.1 - Upper Body Mode Patchwork
I reverted the UE4 control rig: some nodes got disconnected, which maybe signals some data rot?  Very scary.  Anyways, that has been fixed.

For some reason, the metahumans didn't work with upper body mode, even though their control rig and animbp are nearly identical to the ue5 manny which works fine.  I added in some additional logic to help with this, and unfortunately it causes the pelvis to twist 90 degrees.  I can't seem to fix this without effecting other mocap set ups, so I'm leaving it for now.  If anyone has any ideas on why its not working like the other animbps or has an idea for fixing it, feel free to submit a change!

Added a tiny bit of smoothing to the tracker set up used in the movement actor.