UE5.1.1 - Metahumans Rework
Changed the spine solver on the Metahumans, and am getting much better pairing of the head on the body.  I added a similar solver to the UE5 manny face so smooth out spine bends.

Light Rig now has a master intensity, tint and a paramter for blending between your current light set up and the incoming light presets that you load in.

Made a new directory to make it easier to find the full body IK control rigs, and added the up made control rig for post processing, although this has not been implemented yet.