UE5.1.1 - Total Props
Added a requested prop feature to the system!  By swapping out the static mesh in the prop, you can create different items for your characters to use.  You can use buttons in the detail pane to detatch your tracker from the object and reattach it at a different angle, and this offset will be saved for future use in your player saves.

The light rig editor widget can now visually show you the lights in light rig, making it easier for cinematographers to get a sense of what is going on.

The MMVR editor utility widget has been completely re-arranged, and a new function to rotate the player pawn around the players hips has been created.

Added a blueprint I wrote for open XR that doesn't work and makes me sad.

The calibration camera shortcut has been moved to the right thumbstick press (with left thumbstick it was hard to keep the actor from moving around)