UE5.1.1 - Down With OSC (Part 1)
Pretty large changes to the  MMVR player pawn, adding some modularity and tracking solutions for the move to 5.2  There is now a public 'Tracking Method' enum that will spawn in important items on begin play.  Basically attempting to isolate SteamVR to components and blueprints that can later be removed from the system without causing compiler issues on important components.

MMVR now can take in an OSC signal from Brekel OpenVR Recorder.  In Brekel, turn on OSC, set it to localhost on port 7775 with Euler angles and turn on the Brekel tracking method on the player pawn.  This will spawn in yellow gizmos per active tracker.  Calibrating once will determine which tracker is close to the goal you want to drive and create a save game in your project folder called 'BrekelOSC.'  Once this map of trackers to bones is created, calibrating a SECOND time will allow you to drive a character by getting transforms from OSC!

Created a SteamVR hands blueprint that will send finger data to the player pawn so this calculation can be done outside of skeleton specific animation blueprints.

Added a simplified hand rig to each control rig, making it easier for you to clean up animations with increased control over finger data.

Fixed a bug in the CC3 character where it wasn't completing the code on begin play.

Renamed lots of Materials

Fixed a bug in the Quick Select Menu.