UE5.1.1 - Mega Mocap VR OSC Mode
Spent a lot of time working on a way to create OSC communication between two different instance of Unreal Engine!  To use it:

Set up one version of Unreal Engine 5.1 with SteamVR and Live Link XR, and add the IP address of the computer you want to send OSC messages to (there is also a bool on the player pawn that will add a local IP address if you are running your target instance of Unreal on the same computer.)

On your second instance of Unreal that will be receiving the data, change the player pawn's tracking method to 'MMVR_OSC' and put in that computer's IP address (or leave blank if you are transmitting data on the same machine).  Now, tracking data, vr controller input and even MMVR_Props and MMVR_SpectatorCamera tracking will all be sent over the local network!

I have included a new, barebones OSC Map that should be lighter on your system to run, although you can send the data from any player pawn in any level.

Added some new debugging tools to the main editor utility widget.

Revised the UE5 manny, Metahuman and CC3 control rigs, and I think I got rid of the arm popping thanks to a tip from Jeroen Pascal.  Also, it turned out my secondary weight axis on the basic ik node was set to .25 instead of 1, which was causing rotation errors.