UE5.1.1 - Ultra Leap and Foot IK
Added a hand tracking enum on the player pawn that can enable the use of an Ultra Leap hand tracking device.  Changing the enum away from default will override the hand capture method.  Choosing Ultra Leap will spawn in a blueprint that represents the tracking origin for the device, and will attach itself to the chest of the mocap actor target.  You can add this blueprint to your character bluepritn and attach it to any part of your skeletal mesh and this will act as an override location and attachment point, so you can set-up desk, screen top and HMD orientations for the device.

Added a bool on the mocap actor component to allow for Foot IK.  This will try to position the character feet on the ground and offset the pelvis to match, and will attempt to orient the characters feet to the slope of the ground, however this does not work super predictably (at certain foot orientations, there is a snap when the foot rotates on the XY plane.  Any help here would be greatly appreciated!

Set up new foot and hand IK control rig to assist with these features, and added them to every animbp except for IK retargeting just before the final output pose so that they could possibly be used in cinematics.

Cleaned up some of the being play in BP Player pawn.

Tested all skeletons and cleaned the project.