UE5.1.1 - Movement Actor Rig
Added a new MovementActorRig, which is meant to replace the BP_MovementActor.  This rig will get spawned in if the mocap actor component is inside a character blueprint, so you can create your own character logic/have a cleaner set up for gameplay or recording.  I have replaced the movement actor in the MMVR_MovementActorExample level so you can see the new set up, and how it works with the IK foot system.

Noticed a problem with upper body mode, and revised all the control rigs to default to their original bone positions if set to upper body mode.  In testing, I couldn't get default slot/a sequence to combine with the upper body mode: this might need more branching with a blend by bone node in the future.

Changed the way footstep detection works, so it should read and send events even when the character isn't being controlled by a player pawn.

Added a save slot for your player forward offset, so the system will load your forward vector in relation to your pelvis tracker.

Created an event and debugging button  in the Editor Widget for toggling the player pawn visibility.

Added some more camera shake presets for the system.