UE5.2.1 - Look At IK / Anim Slot System
Changed the Mocap Actor component to find a character with the right socket names AND now check if there is a parent that also has the same bone names.  This will allow modular character set ups with child meshes to work with the system.

Fixed a bug on the Spectator Camera that was throwing up errors on the movement actor.

Made the 'create external camera' button the Spectator camera delete the old camera, which should make creating cameras for cutscenes faster.

Fixed the UE4 Manny control rig to preserve scaling.

Added a button for Reseting Root Location that will move the character root to your current center of gravity.

Added a tilt to the movement actor rig based on your speed and direction.

Revised the slot system to include 'face', 'hand_l' and 'hand_r.'  If using these in sequencer, make sure that your clip doesn't have 'Force Custom Mode' enabled under the clip preferences.

Moved the Xbox Controller inputs into the Spectator Camera.  This will allow users to use the spectator camera without needed the MMVR player pawn in the scene as well.

Added a button for foot IK and Look at IK.  Both buttons will look through the currently open Level Sequence to find an actor with a MMVR_MocapActor_Component with a player ID that matches the Editor Utility Widget's Player Int.  Foot IK will force the skeleton to update in editor, and should conform the feet closer to the ground.  Look at IK will spawn in a BP_LookAtTarget if one doesn't exist already in the scene.  The best way to use both modes would be to bake this animation to your rig if it is causing favourable results.  Look at doesn't work with Metahumans because nothing I do can make this work.  Sorry.  Wish I had some help on this one.