UE5.2.1 - Return of the Unknown Actor Profile
I got the Unknown Actor Type working again.  To use it, duplicate the MMVR_Unknown_AnimBP asset and retarget it to whatever skeleton you want to control.  Set up the actor in the typical MMVR fashion, and in the Mocap Actor Component, make sure to chose the profile 'Unknown.'  It would also be helpful to fill out any of the bone names that you want to have be the IK goals in the Actor Profile Bone Names Struct.  When Calibrating, the system will find the closest active tracker to your predetermined bone names, and when no information is provided, the system will create IK goals to the closest valid bone to each remaining tracker.

Changed each IK retargeter to use IK on the feet goals.  If you want to prevent foot sliding, select the feet goals in the chain mapping details panel and change 'Blend To Source' to 1.

Added some text renderers that follow each tracker around during calibration to give insight into which tracker is controlling which VR device.