UE5.2.1 - Auto Finger IK
Added an Auto Finger IK control rig to the UE5, UE4, CC3 and Metahuman Animation Blueprints.  The system will sphere trace along the end finger joints path until it hits an object with visibility.  Unfortunately in my tests this doesn't work with static meshes parented to your character's body inside a blueprint, but does work with objects attached to you character via sequencer.  The Mocap Actor Component controls the Finger IK mode: 0 is off, 1 applies only after animation (not live if you are mocapping) and 2 applies the system after IK (live).

Found a way to get the light rig editor to not stop updating the viewport when it's interacted with.  This should make using it a lot better, although has introduced some bugs.

General light rig usability increases: best feature is if you are working on a cut scene and the light rig editor is open/active in your user interface, the light rig will snap to the camera cuts master camera when using sequencer.

Found the UE5 control rig gave some weird issues in the chest, so added an FBIK node and smoothed the spine slightly with a spline.

Fixed a bug where you couldn't change the focal length on the specator camera when the player pawn wasn't present.  Now the spectator cam should be usable without a player pawn.

The CC3Plus skeleton seemed to have corrupted, so I updated all my Real Illusion plugins and reimported a base character in the project and replaced the existing CC3 skeleton with the new one.  This should allow CC3 characters to be imported again using the skeleton in the project.